# docs/swagger.yaml
openapi : "3.0.0"
info:
  version: 3.5.0
  title: Datasance PoT Controller
paths:
...
  /catalog/microservices:
    get:
      tags:
        - Catalog
      summary: Gets microservices catalog
      operationId: getMicroservicesCatalog
      security:
        - userToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroservicesCatalogResponse"
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
    post:
      tags:
        - Catalog
      summary: Creates a new microservice catalog item
      operationId: createMicroserviceCatalogItem
      security:
        - userToken: []
      requestBody:
        $ref: "#/components/requestBodies/CreateUpdateCatalogItemRequestBody"
      responses:
        "201":
          description: Created
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "409":
          description: Duplicate Name
        "500":
          description: Internal Server Error
  "/catalog/microservices/{id}":
    get:
      tags:
        - Catalog
      summary: Gets microservice catalog item info
      operationId: getMicroserviceCatalogItem
      parameters:
        - in: path
          name: id
          description: Catalog Item Id
          required: true
          schema:
            type: string
      security:
        - userToken: []
      responses:
        "200":
          description: Catalog Item Info
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItemInfoResponse"
        "401":
          description: Not Authorized
        "404":
          description: Invalid Catalog Item Id
        "500":
          description: Internal Server Error
    patch:
      tags:
        - Catalog
      summary: Updates a microservice catalog item
      operationId: updateMicroserviceCatalogItem
      parameters:
        - in: path
          name: id
          description: Catalog Item Id
          required: true
          schema:
            type: string
      security:
        - userToken: []
      requestBody:
        $ref: "#/components/requestBodies/CreateUpdateCatalogItemRequestBody"
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "404":
          description: Invalid Catalog Item Id
        "409":
          description: Duplicate Name
        "500":
          description: Internal Server Error
    delete:
      tags:
        - Catalog
      summary: Deletes a microservice catalog item
      operationId: deleteMicroserviceCatalogItem
      parameters:
        - in: path
          name: id
          description: Catalog Item Id
          required: true
          schema:
            type: string
      security:
        - userToken: []
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "401":
          description: Not Authorized
        "404":
          description: Invalid Catalog Item Id
        "500":
          description: Internal Server Error
...
servers:
  - url: http://localhost:51121/api/v3
components:
  securitySchemes:
    userToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
...
    CreateUpdateCatalogItemRequestBody:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateUpdateCatalogItemRequestBody"
      description: Microservice Catalog Item Info
      required: true
...
  schemas:
...
    MicroservicesCatalogResponse:
      type: object
      properties:
        catalogItems:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItemInfoResponse"
    CatalogItemInfoResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        publisher:
          type: string
        diskRequired:
          type: number
        ramRequired:
          type: number
        picture:
          type: string
        isPublic:
          type: boolean
        registryId:
          type: number
        inputType:
          $ref: "#/components/schemas/InfoTypeResponse"
        outputType:
          $ref: "#/components/schemas/InfoTypeResponse"
        configExample:
          type: string
        images:
          $ref: "#/components/schemas/MicroserviceContainerImages"
    InfoTypeResponse:
      type: object
      properties:
        infoType:
          type: string
        infoFormat:
          type: string
    MicroserviceContainerImages:
      type: array
      items:
        $ref: "#/components/schemas/MicroserviceContainerImage"
    MicroserviceContainerImage:
      type: object
      properties:
        containerImage:
          type: string
        fogTypeId:
          type: number
          enum:
            - 1
            - 2
    CreateUpdateCatalogItemRequestBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        images:
          $ref: "#/components/schemas/MicroserviceContainerImages"
        publisher:
          type: string
        diskRequired:
          type: number
        ramRequired:
          type: number
        picture:
          type: string
        isPublic:
          type: boolean
        registryId:
          type: number
        inputType:
          $ref: "#/components/schemas/InfoTypeResponse"
        outputType:
          $ref: "#/components/schemas/InfoTypeResponse"
        configExample:
          type: string
...